Update to support 8M channels Add additional media types Sync FSEQFile with xLights

+++ src/PixelOverlay.h
@@ -75,7 +75,7 @@ public:
     PixelOverlayModel(FPPChannelMemoryMapControlBlock *block,
                       const std::string &name,
                       char         *chanDataMap,
-                      long long    *pixelMap);
+                      uint32_t     *pixelMap);
     ~PixelOverlayModel();
 
     const std::string &getName() const {return name;};
@@ -123,7 +123,7 @@ private:
     std::string name;
     FPPChannelMemoryMapControlBlock *block;
     char         *chanDataMap;
-    long long    *pixelMap;
+    uint32_t     *pixelMap;
     
     volatile std::thread *updateThread;
     volatile bool threadKeepRunning;
@@ -163,7 +163,7 @@ private:
     std::mutex   modelsLock;
     char         *ctrlMap = nullptr;
     char         *chanDataMap = nullptr;
-    long long    *pixelMap = nullptr;
+    uint32_t     *pixelMap = nullptr;
     
     FPPChannelMemoryMapControlHeader *ctrlHeader = nullptr;
 };
